+++ apps/frontend/src/app/api/openapi.json/route.ts
@@ -1,7 +1,7 @@
-// import { NextResponse } from "next/server";
-// import { openApiDocument } from "~/server/api/openapi";
-//
-// // Respond with our OpenAPI schema
-// export const GET = () => {
-//   return NextResponse.json(openApiDocument);
-// };
+import { NextResponse } from "next/server";
+import { openApiDocument } from "~/server/api/openapi";
+
+// Respond with our OpenAPI schema
+export const GET = () => {
+  return NextResponse.json(openApiDocument);
+};
