+++ apps/frontend/src/server/api/routers/project.ts
@@ -14,11 +14,11 @@ import {
   multiHardwareAccessMiddleware,
 } from "./hardware";
 import { workspaceAccessMiddleware } from "./workspace";
-import { type DatabaseError } from "pg";
 import { type ProjectId, project } from "~/schemas/public/Project";
 import { generateDatabaseId } from "~/lib/id";
 import { markUpdatedAt, getProjectById, getHardwareById } from "~/lib/query";
 import { withDBErrorCheck } from "~/lib/db-utils";
+import { createProject } from "~/server/services/project";
 
 export const projectAccessMiddleware = experimental_standaloneMiddleware<{
   ctx: { db: typeof db; user: { id: string }; workspaceId: string | null };
@@ -77,29 +77,7 @@ export const projectRouter = createTRPCRouter({
         );
 
       return await ctx.db.transaction().execute(async (tx) => {
-        const project = await withDBErrorCheck(
-          tx
-            .insertInto("project")
-            .values({
-              id: generateDatabaseId("project"),
-              ...input,
-            })
-            .returningAll()
-            .executeTakeFirstOrThrow(
-              () =>
-                new TRPCError({
-                  code: "INTERNAL_SERVER_ERROR",
-                  message: "Failed to create project",
-                }),
-            ),
-          {
-            errorCode: "DUPLICATE",
-            errorMsg: `A project with name "${input.name}" for workspace "${ctx.workspace.namespace}" already exists!`,
-          },
-        );
-
-        await markUpdatedAt(tx, "workspace", input.workspaceId);
-        return project;
+        return await createProject(tx, input);
       });
     }),
