+++ apps/frontend/src/server/api/routers/test.ts
@@ -9,8 +9,8 @@ import { insertTestSchema, updateTestSchema } from "~/types/test";
 import { test } from "~/schemas/public/Test";
 import { generateDatabaseId } from "~/lib/id";
 import { markUpdatedAt } from "~/lib/query";
-import { DatabaseError } from "pg";
 import { withDBErrorCheck } from "~/lib/db-utils";
+import { createTest } from "~/server/services/test";
 
 export const testAccessMiddleware = experimental_standaloneMiddleware<{
   ctx: { db: typeof db; user: { id: string }; workspaceId: string | null };
@@ -55,30 +55,7 @@ export const testRouter = createTRPCRouter({
     .output(test)
     .mutation(async ({ ctx, input }) => {
       return await ctx.db.transaction().execute(async (tx) => {
-        const test = await withDBErrorCheck(
-          tx
-            .insertInto("test")
-            .values({
-              id: generateDatabaseId("test"),
-              ...input,
-            })
-            .returningAll()
-            .executeTakeFirstOrThrow(
-              () =>
-                new TRPCError({
-                  code: "INTERNAL_SERVER_ERROR",
-                  message: "Failed to create test",
-                }),
-            ),
-          {
-            errorCode: "DUPLICATE",
-            errorMsg: `A test with name "${input.name}" already exists!`,
-          },
-        );
-
-        await markUpdatedAt(tx, "project", input.projectId);
-
-        return test;
+        return await createTest(tx, input);
       });
     }),
